refactor: split Send-EmailMessage cmdlet
